genericize upgrade banner
