+++ webapp/src/components/rhs/rhs_timeline_placeholder.tsx
@@ -1,214 +1,40 @@
-import React, {useState} from 'react';
-import {useSelector} from 'react-redux';
+// Copyright (c) 2015-present Mattermost, Inc. All Rights Reserved.
+// See LICENSE.txt for license information.
 
+import React from 'react';
 import styled from 'styled-components';
 
-import {getCurrentUser} from 'mattermost-redux/selectors/entities/users';
-import General from 'mattermost-redux/constants/general';
-
-import Spinner from 'src/components/assets/icons/spinner';
-
 import UpgradeTimelineSvg from 'src/components/assets/upgrade_timeline_svg';
+import UpgradeBanner from 'src/components/upgrade_banner';
 import UpgradeTimelineSuccessSvg from 'src/components/assets/upgrade_timeline_success_svg';
 import UpgradeTimelineErrorSvg from 'src/components/assets/upgrade_timeline_error_svg';
 import {PrimaryButton} from 'src/components/assets/buttons';
-import {getAdminAnalytics, isTeamEdition} from 'src/selectors';
-import StartTrialNotice from 'src/components/backstage/start_trial_notice';
-import ConvertEnterpriseNotice from 'src/components/backstage/convert_enterprise_notice';
-
-import {requestTrialLicense, postMessageToAdmins} from 'src/client';
-
 import {AdminNotificationType} from 'src/constants';
-import {isCloud} from 'src/license';
-import {useOpenCloudModal} from 'src/hooks';
-
-enum ActionState {
-    Uninitialized,
-    Loading,
-    Error,
-    Success,
-}
-
-type HandlerType = undefined | (() => (Promise<void> | void));
 
 const TimelineUpgradePlaceholder = () => {
-    const isServerCloud = useSelector(isCloud);
-    const openCloudModal = useOpenCloudModal();
-    const currentUser = useSelector(getCurrentUser);
-    const isCurrentUserAdmin = isSystemAdmin(currentUser.roles);
-    const [actionState, setActionState] = useState(ActionState.Uninitialized);
-    const isServerTeamEdition = useSelector(isTeamEdition);
-
-    const analytics = useSelector(getAdminAnalytics);
-    const serverTotalUsers = analytics?.TOTAL_USERS || 0;
-
-    const endUserMainAction = async () => {
-        if (actionState === ActionState.Loading) {
-            return;
-        }
-
-        setActionState(ActionState.Loading);
-
-        const response = await postMessageToAdmins(AdminNotificationType.VIEW_TIMELINE, isServerTeamEdition);
-        if (response.error) {
-            setActionState(ActionState.Error);
-        } else {
-            setActionState(ActionState.Success);
-        }
-    };
-
-    const requestLicenseSelfHosted = async () => {
-        if (actionState === ActionState.Loading) {
-            return;
-        }
-
-        setActionState(ActionState.Loading);
-
-        const requestedUsers = Math.max(serverTotalUsers, 30);
-        const response = await requestTrialLicense(requestedUsers, AdminNotificationType.VIEW_TIMELINE);
-        if (response.error) {
-            setActionState(ActionState.Error);
-        } else {
-            setActionState(ActionState.Success);
-        }
-    };
-
-    const openUpgradeModal = async () => {
-        if (actionState === ActionState.Loading) {
-            return;
-        }
-
-        openCloudModal();
-    };
-
-    let adminMainAction = requestLicenseSelfHosted;
-    if (isServerCloud) {
-        adminMainAction = openUpgradeModal;
-    }
-
-    let illustration = <UpgradeTimelineSvg/>;
-    let titleText = 'Know what happened';
-    let helpText : React.ReactNode = 'Make retrospectives easy with a timeline that automatically keeps track of the key events and messages so that teams have it at their fingertips.';
-
-    if (isCurrentUserAdmin && isServerTeamEdition) {
-        helpText = <><p>{helpText}</p><ConvertEnterpriseNotice/></>;
-    }
-
-    if (actionState === ActionState.Success) {
-        illustration = <UpgradeTimelineSuccessSvg/>;
-        titleText = 'Thank you!';
-        helpText = 'Your System Admin has been notified.';
-    }
-
-    if (actionState === ActionState.Error) {
-        illustration = <UpgradeTimelineErrorSvg/>;
-        if (isCurrentUserAdmin) {
-            titleText = 'Your license could not be generated';
-            helpText = 'Please check the system logs for more information.';
-        } else {
-            titleText = 'There was an error';
-            helpText = 'We weren\'t able to notify the System Admin.';
-        }
-    }
-
     return (
-        <UpgradeWrapper>
-            {illustration}
-            <UpgradeContent>
-                <UpgradeHeader>
-                    <Title>{titleText}</Title>
-                    <HelpText>{helpText}</HelpText>
-                </UpgradeHeader>
-                <Button
-                    actionState={actionState}
-                    isCurrentUserAdmin={isCurrentUserAdmin}
-                    isServerTeamEdition={isServerTeamEdition}
-                    endUserMainAction={endUserMainAction}
-                    adminMainAction={adminMainAction}
-                    isCloud={isServerCloud}
-                />
-                {!isServerCloud && isCurrentUserAdmin && !isServerTeamEdition && actionState === ActionState.Uninitialized && <Footer/> }
-            </UpgradeContent>
-        </UpgradeWrapper>
+        <UpgradeBanner
+            illustration={<UpgradeTimelineSvg/>}
+            successIllustration={<UpgradeTimelineSuccessSvg/>}
+            errorIllustration={<UpgradeTimelineErrorSvg/>}
+            titleText={'Know what happened'}
+            helpText={'Make retrospectives easy with a timeline that automatically keeps track of the key events and messages so that teams have it at their fingertips.'}
+            notificationType={AdminNotificationType.VIEW_TIMELINE}
+            upgradeWrapperSC={UpgradeWrapper}
+            upgradeContentSC={UpgradeContent}
+            titleSC={Title}
+            helpTextSC={HelpText}
+            buttonSC={PrimaryButton}
+            footerContainerSC={FooterContainer}
+        />
     );
 };
 
-const FooterContainer = styled.div`
-    font-size: 11px;
-    line-height: 16px;
-
+const UpgradeWrapper = styled.div`
     display: flex;
+    flex-direction: column;
     align-items: center;
-    text-align: center;
-
-    color: rgba(var(--center-channel-color, 0.56));
-
-    margin-top: 18px;
 `;
-
-const Footer = () => (
-    <FooterContainer>
-        <StartTrialNotice/>
-    </FooterContainer>
-);
-
-interface ButtonProps {
-    actionState: ActionState;
-    isCurrentUserAdmin: boolean;
-    isServerTeamEdition: boolean;
-    endUserMainAction: HandlerType;
-    adminMainAction: HandlerType;
-    isCloud: boolean;
-}
-
-const Button = (props: ButtonProps) => {
-    if (props.actionState === ActionState.Loading) {
-        return <Spinner/>;
-    }
-
-    if (props.actionState === ActionState.Success) {
-        return null;
-    }
-
-    if (props.actionState === ActionState.Error) {
-        if (props.isCurrentUserAdmin) {
-            return (
-                <PrimaryButton
-                    onClick={() => window.open('https://mattermost.com/support/')}
-                >
-                    {'Contact support'}
-                </PrimaryButton>
-            );
-        }
-        return null;
-    }
-
-    if (props.isCurrentUserAdmin && props.isServerTeamEdition) {
-        return null;
-    }
-
-    let buttonText = 'Notify System Admin';
-    let handleClick : HandlerType = props.endUserMainAction;
-
-    if (props.isCurrentUserAdmin) {
-        handleClick = props.adminMainAction;
-        buttonText = props.isCloud ? 'Upgrade now' : 'Start trial';
-    }
-
-    return (
-        <PrimaryButton
-            onClick={handleClick}
-        >
-            {buttonText}
-        </PrimaryButton>
-    );
-};
-
-const isSystemAdmin = (roles: string): boolean => {
-    const rolesArray = roles.split(' ');
-    return rolesArray.includes(General.SYSTEM_ADMIN_ROLE);
-};
-
 const UpgradeContent = styled.div`
     display: flex;
     flex-direction: column;
@@ -229,7 +55,7 @@ const Title = styled(CenteredRow)`
 
     font-weight: 600;
     font-size: 24px;
-    color: rgba(var(--center-channel-color-rgb), 1);
+    color: var(--center-channel-color);
 `;
 
 const HelpText = styled(CenteredRow)`
@@ -240,14 +66,17 @@ const HelpText = styled(CenteredRow)`
     color: var(--center-channel-color);
 `;
 
-const UpgradeHeader = styled.div`
-    margin-bottom: 14px;
-`;
+const FooterContainer = styled.div`
+    font-size: 11px;
+    line-height: 16px;
 
-const UpgradeWrapper = styled.div`
     display: flex;
-    flex-direction: column;
     align-items: center;
+    text-align: center;
+
+    color: rgba(var(--center-channel-color, 0.56));
+
+    margin-top: 18px;
 `;
 
 export default TimelineUpgradePlaceholder;
